Cleanup duplicate backgrounds
